adding security

+++ src/main/java/ru/mephi/pet/service/GroupService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import ru.mephi.pet.domain.*;
 import ru.mephi.pet.enums.UserACL;
+import ru.mephi.pet.exception.ForbiddenOperationException;
 import ru.mephi.pet.exception.NoACLException;
 import ru.mephi.pet.exception.NotFoundException;
 import ru.mephi.pet.repository.GroupRepository;
@@ -26,6 +27,12 @@ public class GroupService {
     private final UserMapper userMapper;
     private final TaskListMapper taskListMapper;
 
+    private boolean groupHas(Group group, String login) {
+        if (group.getUsers() == null)
+            group = groupRepository.findById(group.getId()).orElseThrow(NotFoundException::new);
+        return group.getUsers().stream().anyMatch(u -> u.getLogin().equals(login));
+    }
+
     public Iterable<GroupDto> getGroups() {
         Iterable<Group> list = groupRepository.findAll();
         List<GroupDto> result = new LinkedList<>();
@@ -34,13 +41,30 @@ public class GroupService {
     }
 
     public GroupDto getGroup(Long id) {
-        return groupMapper.toDto(groupRepository.findById(id).orElseThrow());
+        return groupMapper.toDto(groupRepository.findById(id).orElseThrow(NotFoundException::new));
+    }
+
+    public GroupDto getGroup(String login, Long id) {
+        Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (groupHas(group, login))
+            throw new ForbiddenOperationException();
+        return groupMapper.toDto(group);
     }
 
     public Iterable<UserGroupACLDto> getUsers(Long id) {
         return groupRepository.findById(id)
                 .orElseThrow(NotFoundException::new)
                 .getUserACLS()
+                .stream()
+                .map(userMapper::userGroupACLToUserGroupACLDto)
+                .collect(Collectors.toSet());
+    }
+
+    public Iterable<UserGroupACLDto> getUsers(String login, Long id) {
+        Group group =  groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!groupHas(group, login))
+            throw new ForbiddenOperationException();
+        return group.getUserACLS()
                 .stream()
                 .map(userMapper::userGroupACLToUserGroupACLDto)
                 .collect(Collectors.toSet());
@@ -55,30 +79,49 @@ public class GroupService {
                 .collect(Collectors.toSet());
     }
 
-    public void updateGroup(Long id, GroupDto groupDto) {
+    public Iterable<TaskListDto> getLists(String login, Long id) {
+        Group group =  groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!groupHas(group, login))
+            throw new ForbiddenOperationException();
+        return group.getTasks()
+                .stream()
+                .map(taskListMapper::toDto)
+                .collect(Collectors.toSet());
+    }
+
+    public void updateGroup(String login, Long id, GroupDto groupDto) {
         Group group = groupRepository.findById(id).orElseThrow();
+        if (!findACL(group, login).getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
         group.setName(groupDto.getName());
         groupRepository.save(group);
     }
 
-    public void updateACL(Long id, UserDto userDto, UserACL acl) {
-        Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
-        User user = userMapper.toEntity(userDto);
-        UserGroupACL participant = group
-                .getUserACLS()
+    private UserGroupACL findACL(Group group, String login) {
+        return group.getUserACLS()
                 .stream()
-                .filter(r -> r.getUser().equals(user))
-                .findAny()
-                .orElseThrow(NoACLException::new);
-        if (participant.getUserACL().equals(UserACL.ADMIN))
-            throw new RuntimeException("");
-        participant.setUserACL(acl);
+                .filter(uACL -> uACL.getUser().getLogin().equals(login))
+                .findAny().orElseThrow(NotFoundException::new);
     }
 
-    public void addUser(Long id, UserDto userDto) {
+    public void updateACL(String login, Long id, UserDto userDto, UserACL acl) {
         Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
-        group.getUsers().add(userMapper.toEntity(userDto));
+        User user = userMapper.toEntity(userDto);
+        if (!findACL(group, login).getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
+        UserGroupACL participant = findACL(group, user.getLogin());
+        if (participant.getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
+        participant.setUserACL(acl);
+        userGroupACLRepository.save(participant);
+    }
+
+    public void addUser(String login, Long id, UserDto userDto) {
+        Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!findACL(group, login).getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
         User user = userRepository.findById(userDto.getId()).orElseThrow(NotFoundException::new);
+        group.getUsers().add(user);
         user.getGroups().add(group);
         UserGroupACL userGroupACL = userGroupACLRepository.save(new UserGroupACL());
         userGroupACL.setGroup(group);
@@ -89,8 +132,10 @@ public class GroupService {
         groupRepository.save(group);
     }
 
-    public TaskListDto addList(Long id, TaskListDto list) {
+    public TaskListDto addList(String login, Long id, TaskListDto list) {
         Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (findACL(group, login).getUserACL().equals(UserACL.READER))
+            throw new ForbiddenOperationException();
         TaskList tasks = taskListMapper.toEntity(list);
         tasks.setOwner(group);
         group.getTasks().add(tasks);
@@ -98,17 +143,32 @@ public class GroupService {
         return taskListMapper.toDto(tasks);
     }
 
-    public GroupDto saveGroup(GroupDto group) {
-        return groupMapper.toDto(groupRepository.save(groupMapper.toEntity(group)));
+    public GroupDto saveGroup(String login, GroupDto groupDto) {
+        Group group = groupMapper.toEntity(groupDto);
+        User user = userRepository.findByLogin(login).orElseThrow(NotFoundException::new);
+        group.getUsers().add(user);
+        user.getGroups().add(group);
+        UserGroupACL userGroupACL = userGroupACLRepository.save(new UserGroupACL());
+        userGroupACL.setGroup(group);
+        userGroupACL.setUser(user);
+        userGroupACL.setUserACL(UserACL.ADMIN);
+        group.getUserACLS().add(userGroupACL);
+        user.getGroupACLS().add(userGroupACL);
+        return groupMapper.toDto(groupRepository.save(group));
     }
 
-    public void deleteList(Long id, TaskListDto list) {
+    public void deleteList(String login, Long id, TaskListDto list) {
+        Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!findACL(group, login).getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
         taskListRepository.deleteById(list.getId());
     }
 
-    public void deleteUser(Long id, UserDto userDto) {
-        Group group = groupRepository.findById(id).orElseThrow();
-        User user = userRepository.findById(userDto.getId()).orElseThrow();
+    public void deleteUser(String login, Long id, UserDto userDto) {
+        Group group = groupRepository.findById(id).orElseThrow(NotFoundException::new);
+        if (!findACL(group, login).getUserACL().equals(UserACL.ADMIN))
+            throw new ForbiddenOperationException();
+        User user = userRepository.findById(userDto.getId()).orElseThrow(NotFoundException::new);
         user.getGroups().remove(group);
         group.getUsers().remove(user);
         if (!group.getUserACLS().isEmpty() && group.getUserACLS().stream().noneMatch(g -> g.getUserACL().equals(UserACL.ADMIN)))
